Extract Python session I/O modules
